+++ packages/core/odyssey-design-tokens/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.64.0](https://github.com/atko-eng/odyssey-design-system/compare/v1.63.0...v1.64.0) (2026-07-08)
+
+**Note:** Version bump only for package @okta/odyssey-design-tokens
+
 ## [1.63.0](https://github.com/atko-eng/odyssey-design-system/compare/v1.62.2...v1.63.0) (2026-06-25)
 
 **Note:** Version bump only for package @okta/odyssey-design-tokens